refactor plugin and springboot starter pom
